pyproject.toml
[project]
name = "sandbox"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit>=1.37.1",
    "fastapi[standard]>=0.112.1",
    "uvicorn>=0.30.6",
    "google-api-python-client>=2.142.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
